fix: extract shared env helpers and fix initialized volume icon
