+++ frontend/src/components/k8s/edit-sections/edit-monitoring-section.tsx
@@ -10,30 +10,6 @@ import { KeyValueEditor } from "@/components/common/key-value-editor";
 import type { MonitoringConfig } from "@/lib/api/types";
 
 // ---------------------------------------------------------------------------
-// Helpers: convert between [{name,value}] and flat Record for KeyValueEditor
-// ---------------------------------------------------------------------------
-
-/** Convert [{name: "FOO", value: "bar"}] to {FOO: "bar"} for KeyValueEditor. */
-function envArrayToRecord(
-  arr: Record<string, string>[] | undefined,
-): Record<string, string> | undefined {
-  if (!arr || arr.length === 0) return undefined;
-  const rec: Record<string, string> = {};
-  for (const entry of arr) {
-    if (entry.name) rec[entry.name] = entry.value ?? "";
-  }
-  return Object.keys(rec).length > 0 ? rec : undefined;
-}
-
-/** Convert {FOO: "bar"} to [{name: "FOO", value: "bar"}] for backend format. */
-function recordToEnvArray(
-  rec: Record<string, string> | undefined,
-): Record<string, string>[] | undefined {
-  if (!rec || Object.keys(rec).length === 0) return undefined;
-  return Object.entries(rec).map(([name, value]) => ({ name, value }));
-}
-
-// ---------------------------------------------------------------------------
 // Monitoring Section for Edit Dialog
 // ---------------------------------------------------------------------------
